reorgnize client
